requirements.txt
Django==3.2.10
django-filter==21.1
djangorestframework==3.13.1
flake8==3.8.4
django-rename-app==0.1.3
gunicorn==20.1.0
wheel==0.37.1